+++ util/generate_shell_completions.sh
@@ -12,6 +12,6 @@ cargo build
 spin_path=./target/debug/tspin
 
 # Generate shell completions
-$spin_path --hidden-generate-shell-completions zsh > completions/tspin.zsh
-$spin_path --hidden-generate-shell-completions bash > completions/tspin.bash
-$spin_path --hidden-generate-shell-completions fish > completions/tspin.fish
+$spin_path --generate-zsh-completions > completions/tspin.zsh
+$spin_path --generate-bash-completions > completions/tspin.bash
+$spin_path --generate-fish-completions > completions/tspin.fish